refactor: split pause-point enable resolved-field merge into its own file
pause_point_enable.go exceeded the 500-line production-file cap after the
Expired merge helper was added.
